+++ src/azure-cli/azure/cli/command_modules/acr/commands.py
@@ -24,7 +24,8 @@ from ._format import (
     helm_show_output_format,
     scope_map_output_format,
     token_output_format,
-    token_credential_output_format
+    token_credential_output_format,
+    agentpool_output_format
 )
 from ._client_factory import (
     cf_acr_registries,
@@ -35,7 +36,8 @@ from ._client_factory import (
     cf_acr_runs,
     cf_acr_scope_maps,
     cf_acr_tokens,
-    cf_acr_token_credentials
+    cf_acr_token_credentials,
+    cf_acr_agentpool
 )
 
 
@@ -145,6 +147,12 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
         operations_tmpl='azure.cli.command_modules.acr.token#{}',
         table_transformer=token_credential_output_format,
         client_factory=cf_acr_token_credentials
+    )
+
+    acr_agentpool_util = CliCommandType(
+        operations_tmpl='azure.cli.command_modules.acr.agentpool#{}',
+        table_transformer=agentpool_output_format,
+        client_factory=cf_acr_agentpool
     )
 
     with self.command_group('acr', acr_custom_util) as g:
@@ -294,3 +302,11 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
 
     with self.command_group('acr token credential', acr_token_credential_generate_util) as g:
         g.command('generate', 'acr_token_credential_generate')
+
+    with self.command_group('acr agentpool', acr_agentpool_util, is_preview=True) as g:
+        g.command('create', 'acr_agentpool_create')
+        g.command('update', 'acr_agentpool_update')
+        g.command('delete', 'acr_agentpool_delete')
+        g.command('list', 'acr_agentpool_list')
+        g.command('show', 'acr_agentpool_show')
+        g.command('show-queue', 'acr_agentpool_show_queue')